Revert mock lib into subspaces
