+++ packages/desmos-std/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod mocks;
-
 pub mod subspaces;
 pub mod types;
